AccountHolder and Controller Update

+++ src/main/java/com/meritamerica/capstoneproject/repository/AccountHolderContactDetailsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.meritamerica.capstoneproject.models.AccountHolderContactDetails;
 
-public interface AccountHolderContactDetailsRepository extends JpaRepository<AccountHolderContactDetails, Integer> {
+public interface AccountHolderContactDetailsRepository extends JpaRepository<AccountHolderContactDetails, Long> {
 
 }
