+++ src/main/java/com/meritamerica/capstoneproject/repository/BankAccountRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.meritamerica.capstoneproject.models.BankAccount;
 
-public interface BankAccountRepository extends JpaRepository<BankAccount, Integer> {
+public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
 
 }
